refactor: Tabs 스타일 객체 분리
